requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
channels==4.0.0
channels-redis==4.2.0
python-dotenv==1.0.0
django-allauth==0.58.2
django-filter==23.5
django-storages==1.14.2
Pillow==10.2.0
openai==1.12.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.2.0 